e_book/requirements.txt
django==3.0.3
python-dotenv==0.11.0
django-extensions==2.2.8
cffi==1.14.0
cryptography==2.8
pycparser==2.19
jwt==0.6.1
PyJWT==1.4.0
mysqlclient==1.4.6
xlwt==1.3.0
djangorestframework==3.11.0
Cerberus==1.3.2